Fix DepositWarningModal rename
