+++ src/views/home/components/deposit-warning-modal/deposit-warning-modal.styles.ts
